feat(projects): refactor component imports to follow consistent naming conventions for better readability and maintainability feat(projects): add delete project dialog and empty state components to enhance user experience feat(projects): implement project creation modal and loading skeletons for improved UI feedback during operations feat(projects): add project filters and grid components to facilitate project management and navigation fix(projects): update import paths to match new file naming conventions and ensure proper functionality
