+++ app/(logged-in)/projects/[id]/components/project-content.tsx
@@ -2,10 +2,10 @@
 
 import { useState, useEffect } from 'react';
 
-import ChatInterface from './ChatInterface';
-import CodeExplorer from './CodeExplorer';
-import PreviewPanel from './PreviewPanel';
-import ProjectLayout from './ProjectLayout';
+import ChatInterface from './chat-interface';
+import CodeExplorer from './code-explorer';
+import PreviewPanel from './preview-panel';
+import ProjectLayout from './project-layout';
 import Navbar from '@/components/ui/navbar';
 import { useProjectStore, type Project } from '@/lib/stores/projectStore';
