+++ app/(logged-in)/projects/components/projects-client.tsx
@@ -2,12 +2,12 @@
 
 import { Suspense, useState, useEffect } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
-import ProjectCreationModal from '@/app/(logged-in)/projects/components/ProjectCreationModal';
-import ProjectGrid from '@/app/(logged-in)/projects/components/ProjectGrid';
-import ProjectsHeader from '@/app/(logged-in)/projects/components/ProjectsHeader';
-import EmptyState from './EmptyState';
-import { ProjectsLoadingSkeleton } from './ProjectsLoadingSkeleton';
-import { ProjectModalSkeleton } from './ProjectModalSkeleton';
+import ProjectCreationModal from '@/app/(logged-in)/projects/components/project-creation-modal';
+import ProjectGrid from '@/app/(logged-in)/projects/components/project-grid';
+import ProjectsHeader from '@/app/(logged-in)/projects/components/projects-header';
+import EmptyState from './empty-state';
+import { ProjectsLoadingSkeleton } from './projects-loading-skeleton';
+import { ProjectModalSkeleton } from './project-modal-skeleton';
 import { Project } from '@/lib/stores/projectStore';
 
 interface ProjectsClientProps {
